Move EnumValue to definitions
