+++ polygraph/types/definitions.py
@@ -29,6 +29,20 @@ Field = namedtuple(
 )
 
 
+class EnumValue:
+    __slots__ = ["name", "description", "is_deprecated", "deprecation_reason", "parent"]
+
+    def __init__(self, name, parent, description=None, deprecation_reason=None):
+        self.name = name
+        self.description = description
+        self.is_deprecated = bool(deprecation_reason)
+        self.deprecation_reason = deprecation_reason
+        self.parent = parent
+
+    def __repr__(self):
+        return "EnumValue('{}')".format(self.name)
+
+
 TypeDefinition = namedtuple(
     "TypeDefinition",
     [
